move changes from sink_sql_common to simplify changes
